+++ permissions/src/epilot/models/shared/rolepayload.py
@@ -6,8 +6,12 @@ import dateutil.parser
 from ..shared import grant as shared_grant
 from dataclasses_json import Undefined, dataclass_json
 from datetime import datetime
+from enum import Enum
 from epilot import utils
-from typing import Final, Optional, Union
+from typing import Optional, Union
+
+class RolePayload4Type(str, Enum):
+    PARTNER_ROLE = 'partner_role'
 
 
 @dataclass_json(undefined=Undefined.EXCLUDE)
@@ -25,12 +29,15 @@ class RolePayload4:
     r"""Id of an organization"""
     slug: str = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('slug') }})
     r"""URL-friendly name for the role"""
-    TYPE: Final[str] = dataclasses.field(default='partner_role', metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type') }})
+    type: RolePayload4Type = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type') }})
     expires_at: Optional[datetime] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('expires_at'), 'encoder': utils.datetimeisoformat(True), 'decoder': dateutil.parser.isoparse, 'exclude': lambda f: f is None }})
     r"""date and time then the role will expire"""
     partner_org_id: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('partner_org_id'), 'exclude': lambda f: f is None }})
     
 
+
+class RolePayload3Type(str, Enum):
+    SHARE_ROLE = 'share_role'
 
 
 @dataclass_json(undefined=Undefined.EXCLUDE)
@@ -48,11 +55,14 @@ class RolePayload3:
     r"""Id of an organization"""
     slug: str = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('slug') }})
     r"""URL-friendly name for the role"""
-    TYPE: Final[str] = dataclasses.field(default='share_role', metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type') }})
+    type: RolePayload3Type = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type') }})
     expires_at: Optional[datetime] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('expires_at'), 'encoder': utils.datetimeisoformat(True), 'decoder': dateutil.parser.isoparse, 'exclude': lambda f: f is None }})
     r"""date and time then the role will expire"""
     
 
+
+class RolePayload2Type(str, Enum):
+    ORG_ROLE = 'org_role'
 
 
 @dataclass_json(undefined=Undefined.EXCLUDE)
@@ -70,11 +80,14 @@ class RolePayload2:
     r"""Id of an organization"""
     slug: str = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('slug') }})
     r"""URL-friendly name for the role"""
-    TYPE: Final[str] = dataclasses.field(default='org_role', metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type') }})
+    type: RolePayload2Type = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type') }})
     expires_at: Optional[datetime] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('expires_at'), 'encoder': utils.datetimeisoformat(True), 'decoder': dateutil.parser.isoparse, 'exclude': lambda f: f is None }})
     r"""date and time then the role will expire"""
     
 
+
+class RolePayload1Type(str, Enum):
+    USER_ROLE = 'user_role'
 
 
 @dataclass_json(undefined=Undefined.EXCLUDE)
@@ -92,7 +105,7 @@ class RolePayload1:
     r"""Id of an organization"""
     slug: str = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('slug') }})
     r"""URL-friendly name for the role"""
-    TYPE: Final[str] = dataclasses.field(default='user_role', metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type') }})
+    type: RolePayload1Type = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type') }})
     expires_at: Optional[datetime] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('expires_at'), 'encoder': utils.datetimeisoformat(True), 'decoder': dateutil.parser.isoparse, 'exclude': lambda f: f is None }})
     r"""date and time then the role will expire"""